added user Authentication functionality

+++ controller/expenceController.js
@@ -1,31 +1,33 @@
 const Expence = require('../model/expenceModel');
 
-module.exports.addExpence = (req,res)=>{
-    const {expenceName,amount,description} = req.body;
-    
-    console.log(expenceName,amount,description)
-    Expence.create({
-        expenceName : expenceName,
-        amount : amount,
-        description:description
+module.exports.addExpence = async(req,res)=>{
 
-    })
-    .then((data)=>{
-        console.log(data.toJSON())
-        res.status(200).json({message:"sucssesfully added"});
-    }).catch((err)=>{
-        res.status(400).json({err:"something went wrong"});
-    })
+    try {
+        const {expenceName,amount,description} = req.body;
+        
+       const expence = await Expence.create({ expenceName : expenceName, amount : amount,description:description,userId:req.user.id})
+         return res.status(200).json({message:"sucssesfully added",success:true});
+
+    } catch ( error) {
+        res.status(400).json({err:"something went wrong",success:false});
+    }
+    
 }
 
-module.exports.getExpence = (req,res)=>{
-    Expence.findAll({raw:true}).then((expence)=>{
-        // console.log(expence )
-    return res.status(200).json({expence,sucsses:true});
-        
-    }).catch((err)=>{
-        res.status(500).json(({error:err,success:false}))
-    })
+module.exports.getExpence = async (req,res)=>{
+    
+    try {
+      const expence = await  Expence.findAll({where:{userId:req.user.id}})
+
+      if(!expence) return res.status(400).json({message:"no expence found"});
+
+      return res.status(200).json({expence,sucsses:true});
+
+    } catch (error) {
+        return res.status(500).json({success:false})
+    }
+
+    
 }
 
 
@@ -33,7 +35,8 @@ module.exports.deleteExpence = async(req,res)=>{
 
     try {
        const expenceID = req.params.id;
-       const responce = await Expence.destroy({where:{id:expenceID}});
+       
+       const responce = await Expence.destroy({where:{id:expenceID,userId:req.user.id}});
        res.status(200).json({message:"expence deleted", success:true});
 
     } catch (error) {
